Allow multiple instances at the same time.
Previously, the prototype methods got overwritten every time the
constructor was called. In other words, self did not point to "this",
but rather, pointed to the most recently created instance.

This would have been very funny if anybody tried to use more than one
instance of byline at the same time. But I'm willing to sacrifice the
comedy for the good of mankind.